requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
pillow>=9.0.0
seaborn>=0.12.0
streamlit>=1.35.0
